Another attempt at fixing the setImmediate() issue
It appears the problem is in a polyfill that Webpack automatically
includes when it sees use of the setImmediate() function. This makes
setImmediate() work on a web page, but completely breaks web workers
on load, even if they don't actually call setImmediate().

Trying the slower workaround of setTimeout() in these cases.